Initial Configuration
Created a variety of files that will be necessary for the application.

The index.html page is displaying sample data correctly.

There is now a navigation bar of sorts that offers a variety of options (These options are still yet to be implemented).

Styling is TBD and subject to change.